+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Maintenance
 - Added testing with SSL connection (#1624)
 - Added testing with SSL authentication using CN (#1627)
+- Added testing with TimeSeries NaN values (#1628)
 
 ## v3.3.0 (2025-11-24)
 ### Added